PS_Library_Groups updates
Pretty much works now

+++ deprovision-SustainUser.ps1
@@ -169,18 +169,21 @@ $selectedLeavers | ?{$_.UpnAction -eq "Reassign to another user"} | % {$usersToR
 $sqlConnection = connect-toSqlServer -SQLServer "sql.sustain.co.uk" -SQLDBName "SUSTAIN_LIVE" #This is required to disable ARENA accounts
 #region deprovision
 
-$binMe = convertTo-arrayOfStrings "lucy.boreham@anthesisgroup.com"
+$binMe = convertTo-arrayOfStrings "Matt.Whitehead@anthesisgroup.com,Subhan.Gill@anthesisgroup.com,Kirsty.Smart@anthesisgroup.com"
 foreach($user in $binMe){
     if($user){
         $userMsolObject = Get-User -Identity $user
-        Set-MsolUser -UserPrincipalName $userMsolObject.UserPrincipalName -BlockCredential $true
-        Set-MsolUserPassword -UserPrincipalName $userMsolObject.UserPrincipalName -NewPassword "TTFN123!" -ForceChangePassword $true
-        Get-DistributionGroup -Filter "Members -eq '$($userMsolObject.DistinguishedName)'" | % {
-            Remove-DistributionGroupMember -Identity $_.Id -Member $userMsolObject.UserPrincipalName -Confirm:$false -BypassSecurityGroupManagerCheck:$true
+        if($userMsolObject){
+            write-host "Disabling $($userMsolObject.DisplayName)"
+            Set-MsolUser -UserPrincipalName $userMsolObject.UserPrincipalName -BlockCredential $true
+            Set-MsolUserPassword -UserPrincipalName $userMsolObject.UserPrincipalName -NewPassword "TTFN123!" -ForceChangePassword $true
+            Get-DistributionGroup -Filter "Members -eq '$($userMsolObject.DistinguishedName)'" | % {
+                Remove-DistributionGroupMember -Identity $_.Id -Member $userMsolObject.UserPrincipalName -Confirm:$false -BypassSecurityGroupManagerCheck:$true
+                }
+            Set-Mailbox $userMsolObject.UserPrincipalName -HiddenFromAddressListsEnabled $true -Type Shared
+            Set-MsolUser -UserPrincipalName $userMsolObject.UserPrincipalName -DisplayName $("Ω_"+$userMsolObject.DisplayName) 
+            remove-msolLicenses -userSAM $($userMsolObject.UserPrincipalName.Replace("@anthesisgroup.com",""))
             }
-        Set-Mailbox $userMsolObject.UserPrincipalName -HiddenFromAddressListsEnabled $true -Type Shared
-        Set-MsolUser -UserPrincipalName $userMsolObject.UserPrincipalName -DisplayName $("Ω_"+$userMsolObject.DisplayName) 
-        remove-msolLicenses -userSAM $($userMsolObject.UserPrincipalName.Replace("@anthesisgroup.com",""))
         }
     }
 #-InactiveMailbox
